Use GitHub for theme submodule
While I like SourceHut for its minimalism and no-JavaScript approach,
I'm not keep to promote or support people with fairly blinkered,
black-and-white mentalities who also routinely fail to hold themselves
to the same ethical standards to which they hold others.

It's unfortunate that the actions of such people ironically tend to
have the opposite effect of their stated goals. c.f. RMS and ESR for
other such examples.